+++ partner_telegram/models/__init__.py
@@ -1,3 +1,2 @@
 # License MIT (https://opensource.org/licenses/MIT).
-
-from . import sync_make_module
+from . import res_partner