+++ sync/__manifest__.py
@@ -1,21 +1,23 @@
-# Copyright 2020-2021 Ivan Yelizariev <https://twitter.com/yelizariev>
+# Copyright 2020-2021,2024 Ivan Yelizariev <https://twitter.com/yelizariev>
 # Copyright 2020-2021 Denis Mudarisov <https://github.com/trojikman>
 # Copyright 2021 Ilya Ilchenko <https://github.com/mentalko>
 # License MIT (https://opensource.org/licenses/MIT).
 
 {
-    "name": """Sync Studio""",
-    "summary": """Synchronize something with anything: SystemX↔Odoo, Odoo1↔Odoo2, SystemX↔SystemY. ETL/ESB tool similar to OCA/connector, but more flexible""",
-    "category": "Extra Tools",
-    "images": ["images/sync-studio.jpg"],
-    "version": "16.0.6.2.0",
+    "name": "Sync 🪬 Studio",
+    "summary": """Join the Amazing 😍 Community ⤵️""",
+    "category": "VooDoo ✨ Magic",
+    "version": "16.0.13.0.1",
     "application": True,
-    "author": "IT Projects Labs, Ivan Yelizariev",
-    "support": "help@itpp.dev",
+    "author": "Ivan Kropotkin",
+    "support": "info@odoomagic.com",
     "website": "https://sync_studio.t.me/",
     "license": "Other OSI approved licence",  # MIT
-    "depends": ["base_automation", "mail", "queue_job"],
-    "external_dependencies": {"python": [], "bin": []},
+    # The `partner_telegram` dependency is not directly needed,
+    # but it plays an important role in the **Sync 🪬 Studio** ecosystem
+    # and is added for the quick onboarding of new **Cyber ✨ Pirates**.
+    "depends": ["base_automation", "mail", "queue_job", "partner_telegram"],
+    "external_dependencies": {"python": ["markdown", "pyyaml"], "bin": []},
     "data": [
         "security/sync_groups.xml",
         "security/ir.model.access.csv",
@@ -26,19 +28,19 @@
         "views/sync_trigger_automation_views.xml",
         "views/sync_trigger_webhook_views.xml",
         "views/sync_trigger_button_views.xml",
+        "views/sync_order_views.xml",
         "views/sync_task_views.xml",
-        "views/sync_project_views.xml",
         "views/sync_link_views.xml",
-        "wizard/sync_make_module_views.xml",
+        "views/sync_project_views.xml",
         "data/queue_job_function_data.xml",
     ],
+    "assets": {
+        "web.assets_backend": [
+            "sync/static/src/scss/src.scss",
+        ],
+    },
     "demo": [
-        "data/sync_project_context_demo.xml",
-        "data/sync_project_telegram_demo.xml",
-        "data/sync_project_odoo2odoo_demo.xml",
-        "data/sync_project_trello_github_demo.xml",
         "data/sync_project_unittest_demo.xml",
-        "data/sync_project_context_demo.xml",
     ],
     "qweb": [],
     "post_load": None,