+++ sync/models/sync_task.py
@@ -26,14 +26,20 @@ class SyncTask(models.Model):
     code = fields.Text("Code")
     code_check = fields.Text("Syntax check", store=False, readonly=True)
     active = fields.Boolean(default=True)
+    sync_order_model_id = fields.Many2one("ir.model")
+    sync_order_description = fields.Html(readonly=True)
+    magic_button = fields.Char()
+    button_ids = fields.One2many(
+        "sync.trigger.button", "sync_task_id", string="Manual Triggers", copy=True
+    )
     cron_ids = fields.One2many("sync.trigger.cron", "sync_task_id", copy=True)
     automation_ids = fields.One2many(
         "sync.trigger.automation", "sync_task_id", copy=True
     )
     webhook_ids = fields.One2many("sync.trigger.webhook", "sync_task_id", copy=True)
-    button_ids = fields.One2many(
-        "sync.trigger.button", "sync_task_id", string="Manual Triggers", copy=True
-    )
+    # sync_trigger_order_ids = fields.One2many(
+    #    "sync.trigger.order", "sync_task_id", string="Sync Order Triggers", copy=True
+    # )
     active_cron_ids = fields.Many2many(
         "sync.trigger.cron",
         string="Enabled Crons",
@@ -49,12 +55,6 @@ class SyncTask(models.Model):
     active_webhook_ids = fields.Many2many(
         "sync.trigger.webhook",
         string="Enabled Webhooks",
-        compute="_compute_active_triggers",
-        context={"active_test": False},
-    )
-    active_button_ids = fields.Many2many(
-        "sync.trigger.button",
-        string="Enabled Buttons",
         compute="_compute_active_triggers",
         context={"active_test": False},
     )
@@ -92,14 +92,54 @@ class SyncTask(models.Model):
         "cron_ids.active",
         "automation_ids.active",
         "webhook_ids.active",
-        "button_ids.active",
     )
     def _compute_active_triggers(self):
         for r in self.with_context(active_test=False):
             r.active_cron_ids = r.with_context(active_test=True).cron_ids
             r.active_automation_ids = r.with_context(active_test=True).automation_ids
             r.active_webhook_ids = r.with_context(active_test=True).webhook_ids
-            r.active_button_ids = r.with_context(active_test=True).button_ids
+
+    def action_magic_button(self):
+        # TODO: This should be refactored, because we use single button per task
+        if not self.button_ids:
+            self.button_ids.create(
+                {
+                    "name": self.magic_button,
+                    "trigger_name": self.magic_button,
+                    "sync_task_id": self.id,
+                }
+            )
+        return self.button_ids.start_button()
+
+    def _get_current_date_formatted(self):
+        user_lang = self.env.user.lang or "en_US"
+        lang = self.env["res.lang"].search([("code", "=", user_lang)], limit=1)
+        today = fields.Date.context_today(self)
+        if lang:
+            date_format = lang.date_format
+            formatted_date = today.strftime(date_format)
+        else:
+            formatted_date = today.strftime("%Y-%m-%d")
+
+        return formatted_date
+
+    def action_super_magic_button(self):
+        self.ensure_one()
+        sync_order = self.env["sync.order"].create(
+            {
+                "name": self._get_current_date_formatted(),
+                "sync_task_id": self.id,
+            }
+        )
+        return {
+            "name": "Super 🔥 Magic",
+            "type": "ir.actions.act_window",
+            "view_type": "form",
+            "view_mode": "form",
+            "res_model": "sync.order",
+            "res_id": sync_order.id,
+            "target": "self",
+        }
 
     def start(
         self, trigger, args=None, with_delay=False, force=False, raise_on_error=True
@@ -120,7 +160,7 @@ class SyncTask(models.Model):
         queue_job_or_result = run(
             job, trigger._sync_handler, args, raise_on_error=raise_on_error
         )
-        if with_delay and not self.env.context.get("test_queue_job_no_delay"):
+        if with_delay and not self.env.context.get("queue_job__no_delay"):
             job.queue_job_id = queue_job_or_result.db_record()
             return job
         else: